# server/swagger/swagger.yml
basePath: /api/v2
consumes:
    - application/json
definitions:
    Block:
        description: Block is the basic data unit
        x-go-package: github.com/mattermost/focalboard/server/model
    BlockPatch:
        description: BlockPatch is a patch for modify blocks
        x-go-package: github.com/mattermost/focalboard/server/model
...
    Board:
        description: Board groups a set of blocks and its layout
        x-go-package: github.com/mattermost/focalboard/server/model
...
    BoardPatch:
        description: BoardPatch is a patch for modify boards
        x-go-package: github.com/mattermost/focalboard/server/model
...
    ErrorResponse:
        description: ErrorResponse is an error response
        x-go-package: github.com/mattermost/focalboard/server/model
...
host: localhost
info:
    contact:
        email: api@focalboard.com
        name: Focalboard
        url: https://www.focalboard.com
    description: Focalboard Server
    license:
        name: Custom
        url: https://github.com/mattermost/focalboard/blob/main/LICENSE.txt
    title: Focalboard Server
    version: 2.0.0
paths:
...
    /boards/{boardID}:
        delete:
            description: Removes a board
            operationId: deleteBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        get:
            description: Returns a board
            operationId: getBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        patch:
            description: Partially updates a board
            operationId: patchBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: board patch to apply
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BoardPatch'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
    /boards/{boardID}/archive/export:
        get:
            operationId: archiveExportBoard
            parameters:
                - description: Id of board to export
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
            summary: Exports an archive of all blocks for one boards.
    /boards/{boardID}/blocks:
        get:
            description: Returns blocks
            operationId: getBlocks
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: ID of parent block, omit to specify all blocks
                  in: query
                  name: parent_id
                  type: string
                - description: Type of blocks to return, omit to specify all types
                  in: query
                  name: type
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        items:
                            $ref: '#/definitions/Block'
                        type: array
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        post:
            description: |-
                Insert blocks. The specified IDs will only be used to link
                blocks with existing ones, the rest will be replaced by server
                generated IDs
            operationId: updateBlocks
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: Disables notifications (for bulk inserting)
                  in: query
                  name: disable_notify
                  type: bool
                - description: array of blocks to insert or update
                  in: body
                  name: Body
                  required: true
                  schema:
                    items:
                        $ref: '#/definitions/Block'
                    type: array
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        items:
                            $ref: '#/definitions/Block'
                        type: array
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /boards/{boardID}/blocks/{blockID}:
        delete:
            description: Deletes a block
            operationId: deleteBlock
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: ID of block to delete
                  in: path
                  name: blockID
                  required: true
                  type: string
                - description: Disables notifications (for bulk deletion)
                  in: query
                  name: disable_notify
                  type: bool
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: block not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        patch:
            description: Partially updates a block
            operationId: patchBlock
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: ID of block to patch
                  in: path
                  name: blockID
                  required: true
                  type: string
                - description: Disables notifications (for bulk patching)
                  in: query
                  name: disable_notify
                  type: bool
                - description: block patch to apply
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BlockPatch'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: block not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
    /boards/{boardID}/blocks/{blockID}/duplicate:
        post:
            description: Returns the new created blocks
            operationId: duplicateBlock
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: Block ID
                  in: path
                  name: blockID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        items:
                            $ref: '#/definitions/Block'
                        type: array
                "404":
                    description: board or block not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /teams/{teamID}/archive/export:
        get:
            operationId: archiveExportTeam
            parameters:
                - description: Id of team
                  in: path
                  name: teamID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
            summary: Exports an archive of all blocks for all the boards in a team.
    /teams/{teamID}/archive/import:
        post:
            consumes:
                - multipart/form-data
            operationId: archiveImport
            parameters:
                - description: Team ID
                  in: path
                  name: teamID
                  required: true
                  type: string
                - description: archive file to import
                  in: formData
                  name: file
                  required: true
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: success
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
            summary: Import an archive of boards.
    /boards/{boardID}/archive/import/opportunities:
        post:
            consumes:
                - multipart/form-data
            operationId: importOpportunitiers
            parameters:
                - description: Team ID
                  in: path
                  name: teamID
                  required: true
                  type: string
                - description: xlsx file to import
                  in: formData
                  name: file
                  required: true
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: success
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
            summary: Import opportunities to all boards in a team.
...
produces:
    - application/json
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'Pass session token using Bearer authentication, e.g. set header "Authorization: Bearer <session token>"'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"